+++ link_test.go
@@ -269,7 +269,7 @@ func TestNewSendingLink(t *testing.T) {
 				require.False(t, l.l.dynamicAddr)
 				require.Empty(t, l.l.source.ExpiryPolicy)
 				require.Zero(t, l.l.source.Timeout)
-				require.True(t, l.detachOnDispositionError)
+				require.True(t, l.closeOnDispositionError)
 				require.NotEmpty(t, l.l.key.name)
 				require.Empty(t, l.l.properties)
 				require.Nil(t, l.l.senderSettleMode)
@@ -299,7 +299,7 @@ func TestNewSendingLink(t *testing.T) {
 				require.True(t, l.l.dynamicAddr)
 				require.Equal(t, ExpiryPolicyLinkDetach, l.l.source.ExpiryPolicy)
 				require.Equal(t, uint32(5), l.l.source.Timeout)
-				require.False(t, l.detachOnDispositionError)
+				require.False(t, l.closeOnDispositionError)
 				require.Equal(t, name, l.l.key.name)
 				require.Equal(t, map[encoding.Symbol]any{
 					"property": 123,